[zPIV] Update supply when disconnecting blocks too
