+++ src/zpivchain.h
@@ -5,6 +5,7 @@
 #ifndef PIVX_ZPIVCHAIN_H
 #define PIVX_ZPIVCHAIN_H
 
+#include "chain.h"
 #include "libzerocoin/Coin.h"
 #include "libzerocoin/Denominations.h"
 #include "libzerocoin/CoinSpend.h"
@@ -12,6 +13,7 @@
 #include <string>
 
 class CBlock;
+class CBlockIndex;
 class CBigNum;
 struct CMintMeta;
 class CTransaction;
@@ -40,6 +42,8 @@ std::list<libzerocoin::CoinDenomination> ZerocoinSpendListFromBlock(const CBlock
 /** Global variable for the zerocoin supply */
 extern std::map<libzerocoin::CoinDenomination, int64_t> mapZerocoinSupply;
 int64_t GetZerocoinSupply();
+bool UpdateZPIVSupplyConnect(const CBlock& block, CBlockIndex* pindex, bool fJustCheck);
+bool UpdateZPIVSupplyDisconnect(const CBlock& block, CBlockIndex* pindex);
 
 
 #endif //PIVX_ZPIVCHAIN_H
